@@ -1,6 +1,6 @@
 package top.theillusivec4.diet.common.integration;
 
-import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.screens.Screen;
 import top.theillusivec4.curios.client.gui.CuriosScreen;
 import top.theillusivec4.diet.client.DietClientEventsListener;
 
