@@ -18,9 +18,9 @@
 
 package top.theillusivec4.diet.common.util;
 
-import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.world.entity.player.Player;
 
 public interface PlayerSensitive {
 
-  void setPlayer(PlayerEntity player);
+  void setPlayer(Player player);
 }
