@@ -1,7 +1,7 @@
 package top.theillusivec4.diet.data;
 
 import javax.annotation.Nullable;
-import net.minecraft.data.BlockTagsProvider;
+import net.minecraft.data.tags.BlockTagsProvider;
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import top.theillusivec4.diet.DietMod;
