@@ -1,8 +1,8 @@
 package top.theillusivec4.diet.mixin;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.FoodStats;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.food.FoodData;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -12,13 +12,13 @@ import top.theillusivec4.diet.api.DietCapability;
 import top.theillusivec4.diet.common.util.PlayerSensitive;
 
 @SuppressWarnings("unused")
-@Mixin(FoodStats.class)
-public class MixinFoodStats implements PlayerSensitive {
+@Mixin(FoodData.class)
+public class DietMixinFoodData implements PlayerSensitive {
 
-  PlayerEntity diet_player;
+  Player diet_player;
 
-  @Inject(at = @At("TAIL"), method = "addStats(IF)V")
-  public void diet$addStats(int healing, float saturationModifier, CallbackInfo ci) {
+  @Inject(at = @At("TAIL"), method = "eat(IF)V")
+  public void diet$eat(int healing, float saturationModifier, CallbackInfo ci) {
 
     if (diet_player == null) {
       DietMod.LOGGER.error("Attempted to add food stats to a null player!");
@@ -35,7 +35,7 @@ public class MixinFoodStats implements PlayerSensitive {
   }
 
   @Override
-  public void setPlayer(PlayerEntity playerIn) {
+  public void setPlayer(Player playerIn) {
     diet_player = playerIn;
   }
 }
