@@ -1,7 +1,7 @@
 package top.theillusivec4.diet.mixin;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.FoodStats;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.food.FoodData;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -10,15 +10,15 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import top.theillusivec4.diet.common.util.PlayerSensitive;
 
 @SuppressWarnings("unused")
-@Mixin(PlayerEntity.class)
-public class MixinPlayerEntity {
+@Mixin(Player.class)
+public class DietMixinPlayer {
 
   @Shadow
-  FoodStats foodStats;
+  FoodData foodData;
 
   @SuppressWarnings("ConstantConditions")
   @Inject(at = @At("RETURN"), method = "<init>*")
   public void diet$constructPlayer(CallbackInfo ci) {
-    ((PlayerSensitive) foodStats).setPlayer((PlayerEntity) (Object) this);
+    ((PlayerSensitive) foodData).setPlayer((Player) (Object) this);
   }
 }
